GetBanks becomes unauthenticated call
