+++ build.gradle
@@ -32,5 +32,5 @@ allprojects {
     }
 
     group = 'io.token.sdk'
-    version = '1.0.98'
+    version = '1.0.99'
 }
